Move get_db_version_and_newest_available_version to sqlite.database
